Fully dynamic 16 channel receiver support 1/?
With this commit, we might be officially forced to drop atmega168/328
support because of the lack of resources required for the software to
operate properly (current setup of fully scalable functions for 16
channel receiver requires usage of 64bit variable for aux mask and
function masks.

This will most likely be "too much" for the 2KB of ram available on the
mainstream atmega chips.